+++ Specs/Renderer/AutomaticUniformSpec.js
@@ -2,6 +2,7 @@
 defineSuite([
          'Specs/createContext',
          'Specs/destroyContext',
+         'Specs/createFrameState',
          'Core/Cartesian3',
          'Core/Matrix4',
          'Core/PrimitiveType',
@@ -11,6 +12,7 @@ defineSuite([
      ], 'Renderer/AutomaticUniforms', function(
          createContext,
          destroyContext,
+         createFrameState,
          Cartesian3,
          Matrix4,
          PrimitiveType,
@@ -47,6 +49,9 @@ defineSuite([
                 },
                 getInfiniteProjectionMatrix : function() {
                     return defaultValue(infiniteProjection, Matrix4.IDENTITY.clone());
+                },
+                computeCullingVolume : function() {
+                    return undefined;
                 }
             },
             position : defaultValue(position, Cartesian3.ZERO.clone()),
@@ -154,12 +159,12 @@ defineSuite([
 
     it('has czm_view', function() {
         var us = context.getUniformState();
-        us.update(createMockCamera(
+        us.update(createFrameState(createMockCamera(
             new Matrix4(
                 1.0,  2.0,  3.0,  4.0,
                 5.0,  6.0,  7.0,  8.0,
                 9.0, 10.0, 11.0, 12.0,
-               13.0, 14.0, 15.0, 16.0)), 1.0, new JulianDate());
+               13.0, 14.0, 15.0, 16.0))));
 
         var fs =
             'void main() { ' +
@@ -174,12 +179,12 @@ defineSuite([
 
     it('has czm_viewRotation', function() {
         var us = context.getUniformState();
-        us.update(createMockCamera(
+        us.update(createFrameState(createMockCamera(
             new Matrix4(
                 1.0,  2.0,  3.0,  4.0,
                 5.0,  6.0,  7.0,  8.0,
                 9.0, 10.0, 11.0, 12.0,
-               13.0, 14.0, 15.0, 16.0)), 1.0, new JulianDate());
+               13.0, 14.0, 15.0, 16.0))));
 
         var fs =
             'void main() { ' +
@@ -193,12 +198,12 @@ defineSuite([
 
     it('has czm_inverseView', function() {
         var us = context.getUniformState();
-        us.update(createMockCamera(
+        us.update(createFrameState(createMockCamera(
             new Matrix4(
                0.0, -1.0, 0.0, 7.0,
                1.0,  0.0, 0.0, 8.0,
                0.0,  0.0, 1.0, 0.0,
-               0.0,  0.0, 0.0, 1.0)), 1.0, new JulianDate());
+               0.0,  0.0, 0.0, 1.0))));
 
         var fs =
             'void main() { ' +
@@ -213,12 +218,12 @@ defineSuite([
 
     it('has czm_inverseViewRotation', function() {
         var us = context.getUniformState();
-        us.update(createMockCamera(
+        us.update(createFrameState(createMockCamera(
             new Matrix4(
                0.0, -1.0, 0.0, 7.0,
                1.0,  0.0, 0.0, 8.0,
                0.0,  0.0, 1.0, 9.0,
-               0.0,  0.0, 0.0, 1.0)), 1.0, new JulianDate());
+               0.0,  0.0, 0.0, 1.0))));
 
         var fs =
             'void main() { ' +
@@ -233,13 +238,13 @@ defineSuite([
 
     it('has czm_projection', function() {
         var us = context.getUniformState();
-        us.update(createMockCamera(
+        us.update(createFrameState(createMockCamera(
             undefined,
             new Matrix4(
                1.0,  2.0,  3.0,  4.0,
                5.0,  6.0,  7.0,  8.0,
                9.0, 10.0, 11.0, 12.0,
-              13.0, 14.0, 15.0, 16.0)), 1.0, new JulianDate());
+              13.0, 14.0, 15.0, 16.0))));
 
         var fs =
             'void main() { ' +
@@ -254,13 +259,13 @@ defineSuite([
 
     it('has czm_inverseProjection', function() {
         var us = context.getUniformState();
-        us.update(createMockCamera(
+        us.update(createFrameState(createMockCamera(
             undefined,
             new Matrix4(
                0.0, -1.0, 0.0, 1.0,
                1.0,  0.0, 0.0, 2.0,
                0.0,  0.0, 1.0, 0.0,
-               0.0,  0.0, 0.0, 1.0)), 1.0, new JulianDate());
+               0.0,  0.0, 0.0, 1.0))));
 
         var fs =
             'void main() { ' +
@@ -275,11 +280,11 @@ defineSuite([
 
     it('has czm_infiniteProjection', function() {
         var us = context.getUniformState();
-        us.update(createMockCamera(undefined, undefined,
+        us.update(createFrameState(createMockCamera(undefined, undefined,
             new Matrix4(1.0,  2.0,  3.0,  4.0,
                         5.0,  6.0,  7.0,  8.0,
                         9.0, 10.0, 11.0, 12.0,
-                       13.0, 14.0, 15.0, 16.0)), 1.0, new JulianDate());
+                       13.0, 14.0, 15.0, 16.0))));
 
         var fs =
             'void main() { ' +
@@ -294,12 +299,12 @@ defineSuite([
 
     it('has czm_modelView', function() {
         var us = context.getUniformState();
-        us.update(createMockCamera(
+        us.update(createFrameState(createMockCamera(
             new Matrix4(
                1.0, 0.0, 0.0, 1.0,
                0.0, 1.0, 0.0, 1.0,
                0.0, 0.0, 1.0, 1.0,
-               0.0, 0.0, 0.0, 1.0)), 1.0, new JulianDate());
+               0.0, 0.0, 0.0, 1.0))));
 
         var fs =
             'void main() { ' +
@@ -318,12 +323,12 @@ defineSuite([
 
     it('has czm_modelViewRelativeToEye', function() {
         var us = context.getUniformState();
-        us.update(createMockCamera(
+        us.update(createFrameState(createMockCamera(
             new Matrix4(
                1.0, 0.0, 0.0, 1.0,
                0.0, 1.0, 0.0, 1.0,
                0.0, 0.0, 1.0, 1.0,
-               0.0, 0.0, 0.0, 1.0)), 1.0, new JulianDate());
+               0.0, 0.0, 0.0, 1.0))));
 
         var fs =
             'void main() { ' +
@@ -342,7 +347,7 @@ defineSuite([
 
     it('has czm_inverseModelView', function() {
         var us = context.getUniformState();
-        us.update(createMockCamera(Matrix4.IDENTITY.clone()), 1.0, new JulianDate());
+        us.update(createFrameState(createMockCamera(Matrix4.IDENTITY.clone())));
 
         var fs =
             'void main() { ' +
@@ -361,7 +366,7 @@ defineSuite([
 
     it('has czm_viewProjection', function() {
         var us = context.getUniformState();
-        us.update(createMockCamera(
+        us.update(createFrameState(createMockCamera(
             new Matrix4(1.0, 0.0, 0.0, 0.0,
                         0.0, 1.0, 0.0, 8.0,
                         0.0, 0.0, 1.0, 0.0,
@@ -369,7 +374,7 @@ defineSuite([
             new Matrix4(1.0, 0.0, 0.0, 0.0,
                         0.0, 1.0, 0.0, 0.0,
                         0.0, 0.0, 1.0, 9.0,
-                        0.0, 0.0, 0.0, 1.0)), 1.0, new JulianDate());
+                        0.0, 0.0, 0.0, 1.0))));
 
         var fs =
             'void main() { ' +
@@ -385,7 +390,7 @@ defineSuite([
 
     it('has czm_modelViewProjection', function() {
         var us = context.getUniformState();
-        us.update(createMockCamera(
+        us.update(createFrameState(createMockCamera(
             new Matrix4(1.0, 0.0, 0.0, 0.0,
                         0.0, 1.0, 0.0, 8.0,
                         0.0, 0.0, 1.0, 0.0,
@@ -393,7 +398,7 @@ defineSuite([
             new Matrix4(1.0, 0.0, 0.0, 0.0,
                         0.0, 1.0, 0.0, 0.0,
                         0.0, 0.0, 1.0, 9.0,
-                        0.0, 0.0, 0.0, 1.0)), 1.0, new JulianDate());
+                        0.0, 0.0, 0.0, 1.0))));
 
         var fs =
             'void main() { ' +
@@ -413,7 +418,7 @@ defineSuite([
 
     it('has czm_modelViewProjectionRelativeToEye', function() {
         var us = context.getUniformState();
-        us.update(createMockCamera(
+        us.update(createFrameState(createMockCamera(
             new Matrix4(1.0, 0.0, 0.0, 0.0,
                         0.0, 1.0, 0.0, 8.0,
                         0.0, 0.0, 1.0, 0.0,
@@ -421,7 +426,7 @@ defineSuite([
             new Matrix4(1.0, 0.0, 0.0, 0.0,
                         0.0, 1.0, 0.0, 0.0,
                         0.0, 0.0, 1.0, 9.0,
-                        0.0, 0.0, 0.0, 1.0)), 1.0, new JulianDate());
+                        0.0, 0.0, 0.0, 1.0))));
 
         var fs =
             'void main() { ' +
@@ -441,7 +446,7 @@ defineSuite([
 
     it('has czm_modelViewInfiniteProjection', function() {
         var us = context.getUniformState();
-        us.update(createMockCamera(
+        us.update(createFrameState(createMockCamera(
             new Matrix4(1.0, 0.0, 0.0, 0.0,
                         0.0, 1.0, 0.0, 8.0,
                         0.0, 0.0, 1.0, 0.0,
@@ -450,7 +455,7 @@ defineSuite([
             new Matrix4(1.0, 0.0, 0.0, 0.0,
                         0.0, 1.0, 0.0, 0.0,
                         0.0, 0.0, 1.0, 9.0,
-                        0.0, 0.0, 0.0, 1.0)), 1.0, new JulianDate());
+                        0.0, 0.0, 0.0, 1.0))));
 
         var fs =
             'void main() { ' +
@@ -502,7 +507,7 @@ defineSuite([
 
     it('has czm_encodedCameraPositionMCHigh and czm_encodedCameraPositionMCLow', function() {
         var us = context.getUniformState();
-        us.update(createMockCamera(undefined, undefined, undefined, new Cartesian3(-1000.0, 0.0, 100000.0)), 1.0, new JulianDate());
+        us.update(createFrameState(createMockCamera(undefined, undefined, undefined, new Cartesian3(-1000.0, 0.0, 100000.0))));
 
         var fs =
             'void main() { ' +
@@ -515,25 +520,39 @@ defineSuite([
 
     it('has czm_entireFrustum', function() {
         var us = context.getUniformState();
-        us.update(createMockCamera(), 1.0, new JulianDate());
+        us.update(createFrameState(createMockCamera()));
 
         var fs = 'void main() { gl_FragColor = vec4((czm_entireFrustum.x == 1.0) && (czm_entireFrustum.y == 1000.0)); }';
         verifyDraw(fs);
     });
 
     it('has czm_sunDirectionEC', function() {
+        var us = context.getUniformState();
+        us.update(createFrameState(createMockCamera()));
+
         var fs = 'void main() { gl_FragColor = vec4(czm_sunDirectionEC != vec3(0.0)); }';
         verifyDraw(fs);
     });
 
     it('has czm_sunDirectionWC', function() {
+        var us = context.getUniformState();
+        us.update(createFrameState(createMockCamera()));
+
         var fs = 'void main() { gl_FragColor = vec4(czm_sunDirectionWC != vec3(0.0)); }';
+        verifyDraw(fs);
+    });
+
+    it('has czm_moonDirectionEC', function() {
+        var us = context.getUniformState();
+        us.update(createFrameState(createMockCamera()));
+
+        var fs = 'void main() { gl_FragColor = vec4(czm_moonDirectionEC != vec3(0.0)); }';
         verifyDraw(fs);
     });
 
     it('has czm_viewerPositionWC', function() {
         var us = context.getUniformState();
-        us.update(createMockCamera(), 1.0, new JulianDate());
+        us.update(createFrameState(createMockCamera()));
 
         var fs = 'void main() { gl_FragColor = vec4(czm_viewerPositionWC == vec3(0.0)); }';
         verifyDraw(fs);
@@ -549,8 +568,7 @@ defineSuite([
 
     it('has czm_temeToPseudoFixed', function() {
         var us = context.getUniformState();
-        var time = JulianDate.fromDate(new Date('January 1, 2011 12:00:00 EST'));
-        us.update(createMockCamera(), 1.0, time);
+        us.update(createFrameState(createMockCamera()));
 
         var fs =
             'void main() { ' +