+++ Specs/Scene/BillboardCollectionSpec.js
@@ -4,6 +4,7 @@ defineSuite([
          'Specs/createContext',
          'Specs/destroyContext',
          'Specs/createCamera',
+         'Specs/createFrameState',
          'Specs/frameState',
          'Specs/pick',
          'Specs/render',
@@ -13,7 +14,6 @@ defineSuite([
          'Core/Cartographic',
          'Core/Matrix4',
          'Core/Math',
-         'Core/JulianDate',
          'Renderer/TextureMinificationFilter',
          'Renderer/TextureMagnificationFilter',
          'Renderer/PixelFormat',
@@ -28,6 +28,7 @@ defineSuite([
          createContext,
          destroyContext,
          createCamera,
+         createFrameState,
          frameState,
          pick,
          render,
@@ -37,7 +38,6 @@ defineSuite([
          Cartographic,
          Matrix4,
          CesiumMath,
-         JulianDate,
          TextureMinificationFilter,
          TextureMagnificationFilter,
          PixelFormat,
@@ -70,7 +70,7 @@ defineSuite([
         billboards = new BillboardCollection();
 
         us = context.getUniformState();
-        us.update(createCamera(context), 1.0, JulianDate.fromDate(new Date('January 1, 2011 12:00:00 EST')));
+        us.update(createFrameState(createCamera(context)));
     });
 
     afterEach(function() {