+++ Specs/Scene/LabelCollectionSpec.js
@@ -4,6 +4,7 @@ defineSuite([
          'Specs/createContext',
          'Specs/destroyContext',
          'Specs/createCamera',
+         'Specs/createFrameState',
          'Specs/frameState',
          'Specs/pick',
          'Specs/render',
@@ -14,7 +15,6 @@ defineSuite([
          'Core/Color',
          'Core/Matrix4',
          'Core/Math',
-         'Core/JulianDate',
          'Renderer/BufferUsage',
          'Scene/HorizontalOrigin',
          'Scene/VerticalOrigin',
@@ -26,6 +26,7 @@ defineSuite([
          createContext,
          destroyContext,
          createCamera,
+         createFrameState,
          frameState,
          pick,
          render,
@@ -36,7 +37,6 @@ defineSuite([
          Color,
          Matrix4,
          CesiumMath,
-         JulianDate,
          BufferUsage,
          HorizontalOrigin,
          VerticalOrigin,
@@ -64,7 +64,7 @@ defineSuite([
         labels = new LabelCollection();
 
         us = context.getUniformState();
-        us.update(createCamera(context), 1.0, JulianDate.fromDate(new Date('January 1, 2011 12:00:00 EST')));
+        us.update(createFrameState(createCamera(context)));
     });
 
     afterEach(function() {